doc: fix discrepancy between asciidoc and asciidoctor
Writing a string inside [square brackets], immediately followed
by a string inside `a pair of back quotes`, causes asciidoctor
to eliminate the string inside [square brackets]. Asciidoc is not
affected by this "feature".

Work it around by inserting an otherwise no-op {empty} strings around
the opening and closing square brackets in the verse paragraph of the
synopis to forbid asciidoctor from triggering that feature here.

Unfortunately, this has the side-effect that placeholders in brackets
need to be unconstrained now: [__<placeholder>__]

Signed-off-by: Jean-Noël Avila <jn.avila@free.fr>