xdiff: add new XDL_MERGE_FAVOR_BASE mode
Similar to XDL_MERGE_FAVOR_{OURS,THEIRS} add a new mode for favoring the
base version in the event of a conflict, and add a --base flag to
merge-file to allow testing this out.

This will be used in a subsequent commit to reduce the prevalence of
nested conflict markers which traditionally arise when using the diff3
merge.conflictStyle due to having conflicts in the virtual merge base.
By eschewing using conflicted regions as the resolution for the virtual
merge base, and instead preferring to use the version of the text from
the merge base of the merge bases, we can avoid causing nested conflicts
in the outer merge.

Signed-off-by: Elijah Newren <newren@gmail.com>