xdiff: add new XDL_MERGE_FAVOR_BASE mode
Similar to XDL_MERGE_FAVOR_{OURS,THEIRS} add a new mode for favoring the
base version in the event of a conflict, and add a --base flag to
merge-file to allow testing this out.

This will be used in a subsequent commit to reduce the prevalence of
nested conflict markers in the virtual merge base when using diff3
conflict style, by having the virtual merge base (created by merging the
merge bases) favor the version from their merge base (i.e. from the merge
base of the merge bases).

Signed-off-by: Elijah Newren <newren@gmail.com>