describe,diff: implement --no-optional-locks
Both of these commands may update the index in the background for
similar performance reasons to `git-status`. This commit implements the
`--no-optional-locks` option for these commands, which allows scripts to
bypass this behavior.

I'm implementing this as a solution to a stale lockfile issue we've
sporadically encountered due to a build script that runs `git describe`.
We're mitigating this issue by using `libgit2` in our build script,
which does not have the same background update behavior for its
`git_describe_workdir` implementation, but it would be nice to have this
option supported in the `git` CLI.

Tests and documentation changes are included!

Signed-off-by: Benjamin Woodruff <benjamin.woodruff@vercel.com>